server: Introduce Downloader
Factory Pattern.

So that BackingImage doesn't need to how to actually download or
sync a file. And this abstraction will make the following unit
test introducing possible.

The downloader should be thread safe. Ideally, there won't be a
receive request during pulling, and vice versa. But a deletion
will be invoked regardless of pulling or receiving, we need to
avoid the potential race condition caused by this case.
As for the send function, it shouldn't happen when the backing
image is pulling/receiving files. And there can be multiple
sending simultaneously. Hence adding a check is enough. There is
no need to make sending mutex.

Signed-off-by: Shuo Wu <shuo@rancher.com>