BackingImage does not download URL correctly in some situation
Remove the `Referer` header that is automatically added by the http client. Without this header it is possible to download files that are redirected several times. This mimics the behaviour of `curl` and `wget` which do not submit the `Referer` header by default.

Signed-off-by: Volker Theile <vtheile@suse.com>
(cherry picked from commit df5887369f28b65ce701953d87926627dd78a8c5)