Expose virtual size of qcow2 backing images
This commit adds VirtualSize to the BackingImage, BackingImageStatus,
BackingImageSpec and FileInfo structs. This will allow longhorn-manager
to in turn pick up the virtual image size and expose it in the
status section of the BackingImage and BackingImageManager CRDs.

The virtual size of an image is determined by running `qemu-img info`.
For qcow2 images, the virtual size will usually be larger than the
actual physical file size.  For raw images, `qemu-img info` still reports
virtual size, but it's the same as the physical file size in this case.

It's important to note that we can only report the virtual size of an
image once the syncing file is ready.  If the syncing file is not yet
ready (or if for some reason the call to `qemu-img info` fails), virtual
size will be set to zero.

Related issue: https://github.com/longhorn/longhorn/issues/7923

Signed-off-by: Tim Serong <tserong@suse.com>
(cherry picked from commit ceb51fd0278bbb72f1861f6181b24cd8a467ef55)

# Conflicts:
#	pkg/rpc/rpc.pb.go